// queryeer-desktop/src/plugins/core.editor/texteditor/types.ts
export type TextRange = {
  startLineNumber: number;
  startColumn: number;
  endLineNumber: number;
  endColumn: number;
};